\name{TinnR-package} 
\alias{TinnR-package} 
\docType{package} 
 
\title{ 
  Resources of Tinn-R GUI/Editor for R Environment 
} 
 
\description{ 
  Implements a set of customized functions, adapted from svIDE, svMisc 
  and svIO packages of Philippe Grosjean, necessary to Tinn-R GUI/Editor
  for R environment. 
} 
 
\author{ 
  Adaptations: Jose Claudio Faria \email{joseclaudio.faria@gmail.com}\cr 
  Original author: Philippe Grosjean \email{phgrosjean@sciviews.org} 
} 
 
\references{ 
  \url{URL: http://www.sciviews.org/SciViews-R} 
} 
 
\keyword{package} 
